Add logger, Remove generic from PrivateMemory, Create WorkerFactory
Added a logger than can be used for debug printing to the console.

PrivateMemory is no longer generic, which makes a lot of the classes
that depend on it no longer need to be generic as well. This makes
reflection easier to work with.

Created a WorkerFactory class. This takes a Class object of a Worker
implementation and uses reflection to construct new workers of the
desired type.