Refactor worker thread management, Clean up logger usage in tests
Previously workers were created by wrapping them in Threads and
handling these manually. We now use a ExecutorManager instead and
handle the threads from within Manager. This exposes less thread
management and thread exception handling to the Worker, containing
it in Manager. Additionally, it reduces Thread management overhead
and allows an infinite number of workers since we only ever
use 16 threads to run them.

Added a BeforeAll method to the tests that sets up the logger.