Make memory read type explicit, Finish FoxOtto, Add initialise step
In all use cases of PrivateMemory::read which would previously return
a double by default, we have instead used an explicit readDouble getter.

Added an abstract initialise method to the Worker, which is run once
before he communication and computation phases start.

Finished the implementation of FoxOttoMinPlus and added a holistic
test for this class, which verifies that both the distance and
predecessor matrix produced is correct.