Abstract timing result, Tidy Worker, Make driver use new TimingAnalysis
The result of the computation and communication timing analysis is
stored in a separate TimingAnalysisResult class instead of being stored
in the memory controller.

The Worker methods have been tidied up a bit, and some new utility
methods were added.

The TimedRepeatedMatrixSquaring class was updated to use the new
computation and communication time analysis implemented a few
commits ago.