make GraphCompressor reconstruct paths between deleted nodes
The GraphCompressor can now connect removed nodes up to their
closest neighbour in the compressed graphs and compute the distances
between any two nodes in the original graph correctly.

However, there is still a bug in path reporting in that we still
use both node ids and paths from the original graph for paths between
nodes in the original graph that exists in the compressed one.